Added some example code (sort of)
Added example file to demonstrate each of the individual features. These
were actually the functions I used to test the code. The test functions
are not well documented and are not all that exciting (just some basic
sweeps, switch toggling). Feel free to replace the examples and/or add
better ones.